refactor, address review comments from #3
