+++ src/parser/browser/index.ts
@@ -1,3 +1,3 @@
-import { parse } from './RichTextBrowserParser';
+import { parse } from './rich-text-browser-parser';
 
 export { parse as browserParse };
