+++ src/parser/browser/rich-text-browser-parser.ts
@@ -1,5 +1,5 @@
 import { IDomNode, IOutputResult } from "../parser-models";
-import { convertDomNodeAttributes, getAllNewLineAndWhiteSpace, isElementNode, isRootNode, isTextNode } from "../../utils/";
+import { convertDomNodeAttributes, getAllNewLineAndWhiteSpace, isElementNode, isRootNode, isTextNode } from "../../utils";
 
 export const parse = (input: string): IOutputResult => {
   const document = browserParse(input.replaceAll(getAllNewLineAndWhiteSpace, ''));
