+++ src/parser/node/index.ts
@@ -1,3 +1,3 @@
-import { parse } from './RichTextNodeParser';
+import { parse } from './rich-text-node-parser';
 
 export { parse as nodeParse };
