+++ src/transformers/portable-text-transformer/index.ts
@@ -1,2 +1,1 @@
-export * from './json-transformer';
 export * from './portable-text-transformer';
