--- src/transformers/portable_text_transformer/index.ts
@@ -1,1 +1,0 @@
-export * from './PortableTextTransformer';
