+++ test/transfomers/portable-text-transformer/portable-text-transformer.spec.ts
@@ -154,12 +154,13 @@ describe("portable text transformer", () => {
     expect(result).toMatchSnapshot();
   })
 
-  it("ignores unknown html tags", () => {
+  it("throws error for non-supported tags", () => {
     dummyRichText.value = "<p>text in a paragraph</p><div>text in a div, which doesnt exist in kontent RTE</div>"
     const tree = nodeParse(dummyRichText.value);
-    const result = transformToPortableText(tree);
 
-    expect(result).toMatchSnapshot();
+    expect(() => {
+      transformToPortableText(tree);
+    }).toThrowError();
   })
 
   it("doesn't extend link mark to adjacent spans", () => {
@@ -174,61 +175,7 @@ describe("portable text transformer", () => {
     const tree = nodeParse(dummyRichText.value)
     const result = flatten(tree.children);
 
-    expect(result).toMatchInlineSnapshot(`
-[
-  {
-    "_key": "guid",
-    "_type": "block",
-    "children": [],
-    "markDefs": [],
-    "style": "normal",
-  },
-  {
-    "_key": "guid",
-    "_type": "span",
-    "marks": [],
-    "text": "The coffee drinking culture has been evolving for hundreds and thousands of years. It has never been ",
-  },
-  {
-    "_key": "guid",
-    "_type": "internalLink",
-    "reference": {
-      "_ref": "3120ec15-a4a2-47ec-8ccd-c85ac8ac5ba5",
-      "_type": "reference",
-    },
-  },
-  {
-    "_key": "guid",
-    "_type": "linkMark",
-    "childCount": 2,
-    "value": "guid",
-  },
-  {
-    "_key": "guid",
-    "_type": "span",
-    "marks": [],
-    "text": "so rich as ",
-  },
-  {
-    "_key": "guid",
-    "_type": "mark",
-    "childCount": 1,
-    "value": "strong",
-  },
-  {
-    "_key": "guid",
-    "_type": "span",
-    "marks": [],
-    "text": "today",
-  },
-  {
-    "_key": "guid",
-    "_type": "span",
-    "marks": [],
-    "text": ". How do you make sense of different types of coffee, what do the particular names in beverage menus mean and what beverage to choose for which occasion in your favorite café?",
-  },
-]
-`);
+    expect(result).toMatchSnapshot();
   })
 })
